+++ packages/yourloops/test/integration/assert/remove-patient.assert.ts
@@ -26,22 +26,22 @@
  */
 
 import { act, screen, within } from '@testing-library/react'
-import { monitoredPatient, unmonitoredPatient } from '../data/patient.api.data'
 import userEvent from '@testing-library/user-event'
 import { myThirdTeamId, myThirdTeamName } from '../mock/team.api.mock'
 import PatientApi from '../../../lib/patient/patient.api'
 import DirectShareApi from '../../../lib/share/direct-share.api'
 import { loggedInUserId } from '../mock/auth0.hook.mock'
+import { patient1, patient2 } from '../data/patient.api.data'
 
 export const checkRemovePatientPrivateDialogContent = async () => {
-  const removeButton = screen.getByRole('button', { name: `Remove patient ${monitoredPatient.profile.email}` })
+  const removeButton = screen.getByRole('button', { name: `Remove patient ${patient1.profile.email}` })
   await userEvent.click(removeButton)
   const removeDialog = screen.getByRole('dialog')
 
-  const dialogTitle = within(removeDialog).getByText(`Remove ${monitoredPatient.profile.firstName} ${monitoredPatient.profile.lastName} from My private practice`)
+  const dialogTitle = within(removeDialog).getByText(`Remove ${patient1.profile.firstName} ${patient1.profile.lastName} from My private practice`)
   expect(dialogTitle).toBeVisible()
   const dialogQuestion = within(removeDialog).getByTestId('modal-remove-patient-question')
-  expect(dialogQuestion).toHaveTextContent(`Are you sure you want to remove ${monitoredPatient.profile.firstName} ${monitoredPatient.profile.lastName} from My private practice?`)
+  expect(dialogQuestion).toHaveTextContent(`Are you sure you want to remove ${patient1.profile.firstName} ${patient1.profile.lastName} from My private practice?`)
   const dialogInfo = within(removeDialog).getByText('You will no longer have access to their data.')
   expect(dialogInfo).toBeVisible()
   const confirmRemoveButton = within(removeDialog).getByRole('button', { name: 'Remove patient' })
@@ -53,7 +53,7 @@ export const checkRemovePatientPrivateDialogContent = async () => {
 }
 
 export const checkRemovePatientMedicalTeamDialogContent = async () => {
-  const removeButton = screen.getByRole('button', { name: `Remove patient ${unmonitoredPatient.profile.email}` })
+  const removeButton = screen.getByRole('button', { name: `Remove patient ${patient2.profile.email}` })
   expect(removeButton).toBeVisible()
 
   await userEvent.click(removeButton)
@@ -61,10 +61,10 @@ export const checkRemovePatientMedicalTeamDialogContent = async () => {
   const removeDialog = screen.getByRole('dialog')
   expect(removeDialog).toBeVisible()
 
-  const title = within(removeDialog).getByText(`Remove ${unmonitoredPatient.profile.fullName} from ${myThirdTeamName}`)
+  const title = within(removeDialog).getByText(`Remove ${patient2.profile.fullName} from ${myThirdTeamName}`)
   expect(title).toBeVisible()
   const question = within(removeDialog).getByTestId('modal-remove-patient-question')
-  expect(question).toHaveTextContent(`Are you sure you want to remove ${unmonitoredPatient.profile.fullName} from ${myThirdTeamName}?`)
+  expect(question).toHaveTextContent(`Are you sure you want to remove ${patient2.profile.fullName} from ${myThirdTeamName}?`)
   const info = within(removeDialog).getByText('You and the care team will no longer have access to their data.')
   expect(info).toBeVisible()
   const alertInfo = within(removeDialog).getByText('If you want to remove the patient from another care team, you must first select the care team from the dropdown menu at the top right of YourLoops.')
@@ -76,7 +76,7 @@ export const checkRemovePatientMedicalTeamDialogContent = async () => {
 }
 
 export const checkRemovePatientMedicalTeamConfirm = async () => {
-  const removeButton = screen.getByRole('button', { name: `Remove patient ${unmonitoredPatient.profile.email}` })
+  const removeButton = screen.getByRole('button', { name: `Remove patient ${patient2.profile.email}` })
   expect(removeButton).toBeVisible()
 
   await userEvent.click(removeButton)
@@ -88,14 +88,14 @@ export const checkRemovePatientMedicalTeamConfirm = async () => {
 
   await userEvent.click(confirmRemoveButton)
 
-  expect(PatientApi.removePatient).toHaveBeenCalledWith(myThirdTeamId, unmonitoredPatient.userid)
+  expect(PatientApi.removePatient).toHaveBeenCalledWith(myThirdTeamId, patient2.userid)
   expect(screen.queryByRole('dialog')).not.toBeInTheDocument()
-  expect(screen.getByTestId('alert-snackbar')).toHaveTextContent(`${unmonitoredPatient.profile.firstName} ${unmonitoredPatient.profile.lastName} is no longer a member of ${myThirdTeamName}`)
+  expect(screen.getByTestId('alert-snackbar')).toHaveTextContent(`${patient2.profile.firstName} ${patient2.profile.lastName} is no longer a member of ${myThirdTeamName}`)
 }
 
 export const checkRemovePatientMedicalTeamError = async () => {
   jest.spyOn(PatientApi, 'removePatient').mockRejectedValueOnce(Error('Remove patient error: This error was thrown by a mock on purpose'))
-  const removeButton = screen.getByRole('button', { name: `Remove patient ${monitoredPatient.profile.email}` })
+  const removeButton = screen.getByRole('button', { name: `Remove patient ${patient1.profile.email}` })
   await userEvent.click(removeButton)
   const removeDialog = screen.getByRole('dialog')
   const confirmRemoveButton = within(removeDialog).getByRole('button', { name: 'Remove patient' })
@@ -103,7 +103,7 @@ export const checkRemovePatientMedicalTeamError = async () => {
   await act(async () => {
     await userEvent.click(confirmRemoveButton)
   })
-  expect(PatientApi.removePatient).toHaveBeenCalledWith(myThirdTeamId, monitoredPatient.userid)
+  expect(PatientApi.removePatient).toHaveBeenCalledWith(myThirdTeamId, patient1.userid)
   expect(screen.getByRole('dialog')).toBeVisible()
   expect(screen.getByTestId('alert-snackbar')).toHaveTextContent('Impossible to remove patient. Please try again later.')
   const cancelButton = within(removeDialog).getByRole('button', { name: 'Cancel' })
@@ -113,7 +113,7 @@ export const checkRemovePatientMedicalTeamError = async () => {
 }
 
 export const checkRemovePatientPrivateConfirm = async () => {
-  const removeButton = screen.getByRole('button', { name: `Remove patient ${monitoredPatient.profile.email}` })
+  const removeButton = screen.getByRole('button', { name: `Remove patient ${patient1.profile.email}` })
   await userEvent.click(removeButton)
   const removeDialog = screen.getByRole('dialog')
 
@@ -122,7 +122,7 @@ export const checkRemovePatientPrivateConfirm = async () => {
 
   await userEvent.click(confirmRemoveButton)
 
-  expect(DirectShareApi.removeDirectShare).toHaveBeenCalledWith(monitoredPatient.userid, loggedInUserId)
+  expect(DirectShareApi.removeDirectShare).toHaveBeenCalledWith(patient1.userid, loggedInUserId)
   expect(screen.queryByRole('dialog')).not.toBeInTheDocument()
-  expect(screen.getByTestId('alert-snackbar')).toHaveTextContent(`Direct data sharing with ${monitoredPatient.profile.firstName} ${monitoredPatient.profile.lastName} has been removed`)
+  expect(screen.getByTestId('alert-snackbar')).toHaveTextContent(`Direct data sharing with ${patient1.profile.firstName} ${patient1.profile.lastName} has been removed`)
 }