+++ CHANGELOG.md
@@ -13,6 +13,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2154 Displayed columns choice in the patient list
 - YLP-2160 Patient list new design
 - YLP-2172 Display patients count on patients list
+- YLP-2173 As a HCP, I want information when hovering the monitoring alerts icons
 - YLP-2175 Implement new design system
 - YLP-2182 Sort patients by number of monitoring alerts on patient list
 - YLP-2183 Merging monitoring alerts into 1 column in the patient list (step 1)
@@ -27,6 +28,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2220 Remove "private practice" filter from patients list
 - YLP-2231 Implement new medical reports design for patients
 - YLP-2265 Disable Add patient button when private practice is selected
+- YLP-2312 Teams and patients in a team should always have a field "monitoringAlertsParameters"
 - YLP-2322 Pending patients should not display glycemia specific data
 - YLP-2323 Add new columns from existing data in patient list
 - YLP-2324 Hide certain columns when private practice is selected